refactor : (Auth) DDD Layer에 맞는 위치 변경
