+++ orury-domain/src/main/java/org/orury/domain/auth/domain/dto/LoginDto.java
@@ -1,4 +1,4 @@
-package org.orury.domain.auth.dto;
+package org.orury.domain.auth.domain.dto;
 
 import org.orury.domain.global.constants.NumberConstants;
 import org.orury.domain.user.domain.dto.UserDto;
