+++ orury-domain/src/main/java/org/orury/domain/auth/domain/dto/apple/AppleIdTokenPayload.java
@@ -1,4 +1,4 @@
-package org.orury.domain.auth.dto.apple;
+package org.orury.domain.auth.domain.dto.apple;
 
 public record AppleIdTokenPayload(
         String sub,
