+++ orury-domain/src/main/java/org/orury/domain/auth/domain/dto/kakao/KakaoAccountDto.java
@@ -1,4 +1,4 @@
-package org.orury.domain.auth.dto.kakao;
+package org.orury.domain.auth.domain.dto.kakao;
 
 import java.time.LocalDateTime;
