+++ orury-domain/src/main/java/org/orury/domain/auth/domain/dto/kakao/KakaoOAuthTokenDto.java
@@ -1,4 +1,4 @@
-package org.orury.domain.auth.dto.kakao;
+package org.orury.domain.auth.domain.dto.kakao;
 
 public record KakaoOAuthTokenDto(
         String accessToken,
