+++ orury-domain/src/main/java/org/orury/domain/auth/infrastructure/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
-package org.orury.domain.auth.db.repository;
+package org.orury.domain.auth.infrastructure;
 
-import org.orury.domain.auth.db.model.RefreshToken;
+import org.orury.domain.auth.domain.entity.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
